Split up the Internal module
