+++ tasty-plutus/tasty-plutus.cabal
@@ -55,6 +55,10 @@ library
 
   other-modules:
     Test.Tasty.Plutus.Internal
+    Test.Tasty.Plutus.Internal.Context
+    Test.Tasty.Plutus.Internal.Options
+    Test.Tasty.Plutus.Internal.Run
+    Test.Tasty.Plutus.Internal.WithScript
     Test.Tasty.Plutus.Options.QQ
 
   build-depends:
